refactor: place Next entry base client references
