+++ docs/nextjs-fidelity-architecture-tracker.md
@@ -283,8 +283,8 @@ higher owner is blocked.
 | --------------------------------------------- | ----------------------------------------------------------------------------------------- | ------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `plugin/builtin-global-error.ts`              | `src/client/components/builtin/global-error.ts`                                           | Builtin global error component/reference behavior | Done in this worktree. Vite virtual module preserves the built-in client reference shape; `.ts` is used because the plugin graph is loaded by Vitest's native config loader. P1 deletion target: let real Next app entries provide the built-in global-error module. |
 | `plugin/builtin-global-error.test.ts`         | `src/client/components/builtin/global-error.test.ts`                                      | Same                                              | Done in this worktree. Test follows implementation.                                                                                                                                                                                                                  |
-| `plugin/entry-base-client-references.ts`      | `src/server/app-render/entry-base.ts`                                                     | Next app-render entry-base client refs            | Move mirror behavior there.                                                                                                                                                                                                                                          |
-| `plugin/entry-base-client-references.test.ts` | `src/server/app-render/entry-base.test.ts`                                                | Same                                              | Test follows implementation.                                                                                                                                                                                                                                         |
+| `plugin/entry-base-client-references.ts`      | `src/server/app-render/entry-base.ts`                                                     | Next app-render entry-base client refs            | Done in this worktree. Vite bridge preserves Next entry-base client reference boundaries. P1 deletion target: let `@vitejs/plugin-rsc` preserve CJS `"use client"` dependency boundaries during RSC dep optimization.                                                |
+| `plugin/entry-base-client-references.test.ts` | `src/server/app-render/entry-base.test.ts`                                                | Same                                              | Done in this worktree. Test follows implementation.                                                                                                                                                                                                                  |
 | `plugin/server-reference-info.ts`             | `src/shared/lib/server-reference-info.ts`                                                 | Next server reference info helper                 | Move mirror behavior there.                                                                                                                                                                                                                                          |
 | `plugin/server-reference-info.test.ts`        | `src/shared/lib/server-reference-info.test.ts`                                            | Same                                              | Test follows implementation.                                                                                                                                                                                                                                         |
 | `plugin/runtime-rewrites.ts`                  | `src/client/app-index.tsx` or `src/server/web/adapter.ts` depending on content            | Runtime/browser bootstrap rewrites                | Inspect content before move; must name exact source.                                                                                                                                                                                                                 |
